Relicense to MIT
There's no big news here. I have just been reading more about licensing
and have decided that I would prefer to share this software under the
more permissive terms of the MIT license. I also think that this
software is simple enough that anyone who wanted to publish a derivative
work but did not like the terms of the GPL could easily rewrite the
software and license it differently.